pyproject.toml
[project]
name = "algotrading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.29.5",
    "kiteconnect>=5.0.1",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
]
